Move JobQueue to be a module so that we can use it in preload
