+++ test/index.html
@@ -373,7 +373,6 @@
   <script type='text/javascript' src='../js/expiring_messages.js' data-cover></script>
   <script type='text/javascript' src='../js/notifications.js' data-cover></script>
   <script type='text/javascript' src='../js/focus_listener.js'></script>
-  <script type='text/javascript' src='../js/job_queue.js'></script>
 
   <script type="text/javascript" src="../js/chromium.js" data-cover></script>
